requirements.txt
absl-py 
attrs
certifi 
cffi
charset==normalizer
colorama 
contourpy 
cycler
filelock 
flatbuffers 
fonttools
fsspec 
huggingface==hub 
idna 
imageio 
importlib-metadata 
importlib-resources
intel-openmp 
jax 
jaxlib 
jinja2 
joblib 
kiwisolver 
lazy-loader 
markupsafe 
matplotlib 
mediapipe 
mkl 
ml-dtypes 
mpmath 
networkx 
numpy 
opencv-contrib-python 
opencv-python
opt-einsum 
packaging 
pandas 
pick 
pillow 
protobuf 
pycparser 
pyparsing 
python-dateutil
pytz
pyyaml 
requests 
safetensors 
scikit-image 
scikit-learn 
scipy 
sentencepiece 
six 
sounddevice 
sympy 
tbb 
threadpoolctl 
tifffile 
timm 
torch 
torchaudio 
torchvision 
tqdm
typing-extensions
tzdata 
urllib3
windows-curses 
zipp 